+++ src/main/php/web/Routing.class.php
@@ -11,7 +11,6 @@ use web\routing\{CannotRoute, Path, Target};
  * @test  web.unittest.RoutingTest
  */
 class Routing implements Handler {
-  private $top= false;
   private $fallback= null;
   private $routes= [];
 
@@ -23,12 +22,11 @@ class Routing implements Handler {
    * - A handler, which becomes the argument to `fallback()`.
    *
    * @param  web.Handler|web.Application|function(web.Request, web.Response): var|[:var] $routes
-   * @param  bool $top Whether this is the top-level routing
    * @return self
    */
-  public static function cast($routes, $top= false) {
+  public static function cast($routes) {
     if ($routes instanceof self) {
-      $r= $routes;
+      return $routes;
     } else if ($routes instanceof Application) {
       $r= new self();
       foreach ($routes->routing()->routes() as $pattern => $target) {
@@ -37,17 +35,16 @@ class Routing implements Handler {
         };
         $r->with(new Route($matcher, $target));
       }
+      return $r;
     } else if (is_array($routes)) {
       $r= new self();
       foreach ($routes as $definition => $target) {
         $r->matching($definition, $target);
       }
+      return $r;
     } else {
-      $r= (new self())->fallbacks($routes);
+      return (new self())->fallbacks($routes);
     }
-
-    $r->top= $top;
-    return $r;
   }
 
   /** @return web.routing.Route[] */
@@ -147,23 +144,11 @@ class Routing implements Handler {
    * @return var
    */
   public function handle($request, $response) {
-    $seen= [];
-
-    dispatch: $result= $this->route($request)->handle($request, $response);
-    if ($this->top && $result instanceof Dispatch) {
-      $seen[$request->uri()->hashCode()]= true;
-      $request->rewrite($result->uri());
-      if (isset($seen[$request->uri()->hashCode()])) {
-        throw new Error(508, 'Internal redirect loop caused by dispatch to '.$result->uri());
-      }
-      goto dispatch;
-    }
-
-    return $result;
+    return $this->route($request)->handle($request, $response);
   }
 
   /** @deprecated */
   public function service($request, $response) {
-    return $this->handle($request, $response);
+    return $this->route($request)->handle($request, $response);
   }
 }